# cli/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tools API",
    "description": "\nInternal APIs for GlueOps platform engineers: stand up dev/test infrastructure,\ngenerate cluster manifests, and tear it all down when you are done.\n\n### Conventions\n\n* Most endpoints are keyed on a **`captain_domain`** (e.g. `nonprod.foobar.onglueops.rocks`).\n  The first label (`nonprod`) is typically the tenant namespace.\n* Manifest endpoints return **plain-text YAML** that you commit to a deployment-configurations\n  repository \u2014 they do not apply anything to a cluster.\n* Every example value in this page is a real, working example. Click **Try it out** on any\n  endpoint to fire the request straight from here.\n\n### \u26a0\ufe0f Destructive endpoints\n\nOperations marked **(destructive)** delete real infrastructure, repositories, or backups.\nDouble-check the `captain_domain` / account name before running them.\n\n### Other views\n\n[ReDoc](/redoc) \u00b7 [OpenAPI spec](/openapi.json) \u00b7\nthe [`tools` CLI](https://github.com/GlueOps/tools-api) wraps every endpoint below for headless machines.\n",
    "version": "UNKNOWN"
  },
  "paths": {
    "/v1/storage-buckets": {
      "post": {
        "tags": [
          "Storage & Backups"
        ],
        "summary": "Recreate monitoring storage buckets (destructive)",
        "description": "Create/re-create the storage buckets used by V2 of our monitoring stack (the Otel based one).\n\nNote: this can be a DESTRUCTIVE operation.\nFor the provided captain_domain, this will DELETE and then create new/empty storage buckets for loki, tempo, and thanos.",
        "operationId": "hello_v1_storage_buckets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageBucketsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/setup-aws-account-credentials": {
      "post": {
        "tags": [
          "AWS"
        ],
        "summary": "Get admin credentials for an AWS sub-account",
        "description": "If you are testing in AWS/EKS you will need an AWS account to test with. This request will provide you with admin level credentials to the sub account you specify.\nThis can also be used to just get Admin access to a desired sub account.",
        "operationId": "create_credentials_for_aws_captain_account_v1_setup_aws_account_credentials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AwsCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nuke-aws-captain-account": {
      "delete": {
        "tags": [
          "AWS"
        ],
        "summary": "Nuke an AWS sub-account (destructive)",
        "description": "Run this after you are done testing within AWS. This will clean up orphaned resources.\n\nSubmit the AWS account name you want to nuke (e.g. glueops-captain-foobar).\n\nNote: you may have to run this 2x.",
        "operationId": "nuke_aws_captain_account_v1_nuke_aws_captain_account_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AwsNukeAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nuke-captain-domain-data": {
      "delete": {
        "tags": [
          "Storage & Backups"
        ],
        "summary": "Delete all backups/data for a captain domain (destructive)",
        "description": "Deletes all backups/data for a provided captain_domain. Running this before a cluster creation\nhelps ensure a clean environment.\n\nSubmit the captain_domain/tenant you want to nuke (e.g. nonprod.foobar.onglueops.rocks). This will delete all backups and data for the provided captain_domain.\n\nThis will remove things like the vault and cert-manager backups.\n\nNote: this may not delete things like Loki/Thanos/Tempo data as that may be managed outside of AWS.",
        "operationId": "nuke_captain_domain_data_v1_nuke_captain_domain_data_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptainDomainNukeDataAndBackupsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reset-github-organization": {
      "delete": {
        "tags": [
          "GitHub"
        ],
        "summary": "Reset a tenant GitHub organization (destructive)",
        "description": "Resets the GitHub Organization to make it easier to get a new dev cluster running for Dev.\n\nSubmit the dev captain_domain you want to nuke (e.g. nonprod.foobar.onglueops.rocks). This will reset the GitHub organization so that you can easily get up and running with a new dev cluster.\n\nThis will reset your deployment-configurations repository, it'll bring over a working regcred, and application repos with working github actions so that you can quickly work on the GlueOps stack.\n\nWARNING: By default delete_all_existing_repos = True. Please set it to False or make a manual backup if you are concerned about any data loss within your tenant org (e.g. github.com/development-tenant-*)",
        "operationId": "reset_github_organization_v1_reset_github_organization_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetGitHubOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/github/workflow-run-status": {
      "post": {
        "tags": [
          "GitHub"
        ],
        "summary": "Get the status of a GitHub Actions workflow run",
        "description": "Provide a GitHub Actions run URL (e.g. https://github.com/owner/repo/actions/runs/12345678) and get the current status of that workflow run.\nWorks for any repo the configured GITHUB_TOKEN has read access to.",
        "operationId": "get_workflow_run_status_v1_github_workflow_run_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubWorkflowRunStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/k3d-lb-nodes": {
      "post": {
        "tags": [
          "Load Balancers"
        ],
        "summary": "Create k3d-lb nodes on Proxmox",
        "description": "Creates Chisel nodes on Proxmox (via Waggle placement) for dev/k3d clusters. This allows us to\nmimic a Cloud Controller for Loadbalancers (e.g. NLBs with EKS).\n\nIf you are testing within k3ds you will need chisel to provide you with load balancers.\nFor a provided captain_domain this will delete any existing k3d-lb nodes and provision new ones.\nPlacement is decided by Waggle (pool per captain_domain); the VMs are then created on the assigned\nProxmox hypervisors and their IPs are read back through the QEMU guest agent.\nNote: this will generally result in new IPs being provisioned.",
        "operationId": "create_k3d_lb_nodes_v1_k3d_lb_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/K3dLbNodesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Load Balancers"
        ],
        "summary": "Delete k3d-lb nodes on Proxmox",
        "description": "Deletes your k3d-lb nodes. Please run this when you are done with development to free up capacity.\n\nWhen you are done testing with k3ds this will delete your k3d-lb nodes (Proxmox VMs + Waggle pool) and free up capacity.",
        "operationId": "delete_k3d_lb_nodes_v1_k3d_lb_nodes_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/K3dLbNodesDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/opsgenie": {
      "post": {
        "tags": [
          "Alerting"
        ],
        "summary": "Generate Opsgenie alerts manifest",
        "description": "Create a opsgenie/alertmanager configuration. Do this for any clusters you want alerts on.",
        "operationId": "create_opsgeniealerts_manifest_v1_opsgenie_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsgenieAlertsManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidentio": {
      "post": {
        "tags": [
          "Alerting"
        ],
        "summary": "Generate incident.io alerts manifest",
        "description": "Create an incident.io/alertmanager configuration. Do this for any clusters you want alerts on.",
        "operationId": "create_incidentioalerts_manifest_v1_incidentio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentioAlertsManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kube-apiserver": {
      "post": {
        "tags": [
          "Manifests"
        ],
        "summary": "Generate kube-apiserver exposure manifest",
        "description": "Expose the cluster kube-apiserver via Traefik (TLS passthrough + IP allowlist).\n\nGenerate the Namespace + Traefik MiddlewareTCP + IngressRouteTCP manifest that\nexposes the cluster's Kubernetes API server at kube-api.<captain_domain>,\nrestricted to the provided IP allowlist, with TLS passthrough.\n\nCluster prerequisite: the IngressRouteTCP references the kubernetes Service in the\ndefault namespace from glueops-core-kube-api, so the platform Traefik must have\nproviders.kubernetesCRD.allowCrossNamespace=true or the route is silently dropped.",
        "operationId": "create_kube_apiserver_manifest_v1_kube_apiserver_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KubeApiserverManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kube-rbac": {
      "post": {
        "tags": [
          "Manifests"
        ],
        "summary": "Generate developer-debug RBAC manifest",
        "description": "Developer-debug RBAC (reader/reader-plus/debugger/operator) for a tenant's namespace.\n\nGenerate the ClusterRoles + namespace-scoped RoleBindings that let a tenant's developers\ndebug their workloads (Lens/k9s) in their <environment> namespace via the kube-apiserver\nexposed by /v1/kube-apiserver. The namespace is the first label of captain_domain and the\nRoleBinding subjects are oidc:<tenant_github_organization_name>:<captain_domain>-kubectl-<role>.\nAlso includes the hardcoded glueops-super-admins -> cluster-admin ClusterRoleBinding.",
        "operationId": "create_kube_rbac_manifest_v1_kube_rbac_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KubeRbacManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/captain-manifests": {
      "post": {
        "tags": [
          "Manifests"
        ],
        "summary": "Generate captain manifests",
        "description": "Generate YAML manifests for captain deployments based on the provided configuration.",
        "operationId": "create_captain_manifests_v1_captain_manifests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptainManifestsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AwsCredentialsRequest": {
        "properties": {
          "aws_sub_account_name": {
            "type": "string",
            "title": "Aws Sub Account Name",
            "example": "glueops-captain-foobar"
          }
        },
        "type": "object",
        "required": [
          "aws_sub_account_name"
        ],
        "title": "AwsCredentialsRequest"
      },
      "AwsNukeAccountRequest": {
        "properties": {
          "aws_sub_account_name": {
            "type": "string",
            "title": "Aws Sub Account Name",
            "example": "glueops-captain-foobar"
          }
        },
        "type": "object",
        "required": [
          "aws_sub_account_name"
        ],
        "title": "AwsNukeAccountRequest"
      },
      "CaptainDomainNukeDataAndBackupsRequest": {
        "properties": {
          "captain_domain": {
            "type": "string",
            "title": "Captain Domain",
            "example": "nonprod.foobar.onglueops.rocks"
          }
        },
        "type": "object",
        "required": [
          "captain_domain"
        ],
        "title": "CaptainDomainNukeDataAndBackupsRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StorageBucketsRequest": {
        "properties": {
          "captain_domain": {
            "type": "string",
            "title": "Captain Domain",
            "example": "nonprod.foobar.onglueops.rocks"
          }
        },
        "type": "object",
        "required": [
          "captain_domain"
        ],
        "title": "StorageBucketsRequest"
      },
...
    }
  },
...
}